Screen-specific scaffolds
We used to have a single scaffold for the entire app with the home screen and the list screen showing the content area of the scaffold. This leads to navigation problems, particularly related to how the top bar shows. The expected usage is for each screen to use its own scaffold.